server/requirements.txt
Flask
gunicorn
pyjwt
SQLAlchemy
mysqlclient
